+++ app/Services/LocationService.php
@@ -32,6 +32,7 @@ class LocationService
             'company' => 'locations.company',
             'city' => 'locations.city',
             'type' => 'location_types.title',
+            'estimated_waiting_time' => 'locations.estimated_waiting_time',
         ];
         $sortBy = (string) ($filters['sort_by'] ?? 'created_at');
         $sortColumn = $sortableColumns[$sortBy] ?? $sortableColumns['created_at'];
@@ -202,6 +203,7 @@ class LocationService
             'last_name' => $data['last_name'] ?? null,
             'phone' => $data['phone'] ?? null,
             'email' => $data['email'] ?? null,
+            'estimated_waiting_time' => $data['estimated_waiting_time'] ?? null,
             'province' => $data['province'],
             'post_code' => $data['post_code'],
             'latitude' => $data['latitude'] ?? null,
@@ -262,6 +264,7 @@ class LocationService
             'last_name',
             'phone',
             'email',
+            'estimated_waiting_time',
             'province',
             'post_code',
             'latitude',
@@ -303,6 +306,7 @@ class LocationService
             'last_name',
             'phone',
             'email',
+            'estimated_waiting_time',
             'province',
             'post_code',
             'latitude',